Update: minor backend refactor + move of ReadMe
