+++ rust-warp-jwt-example/common/src/lib.rs
@@ -1,7 +1,26 @@
+use serde::{Deserialize, Serialize};
+
 #[derive(Clone)]
 pub struct User{
     pub uid: String,
     pub email: String,
     pub pw: String,
     pub role: String,
+    // TODO: potentially to be extended
+}
+
+
+
+#[derive(Deserialize)]
+pub struct LoginRequest{
+    pub email: String,
+    pub pw: String
+}
+
+#[derive(Serialize)]
+pub struct LoginResponse{
+    // This (JWT) token is returned to the client, so the user in the future, 
+    // can use it to further (then authenticated) requests by attaching it to the
+    // header.
+    pub token: String
 }
